refactor: isolate production deploy gate
